% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMaximumQCM.R
\name{localMaximumQCM}
\alias{localMaximumQCM}
\title{localMaximumQCM: Subroutine for Creating Gene Clusters}
\usage{
localMaximumQCM(cMatrix, gamma = 0.55, t = 1, lambda = 1)
}
\arguments{
\item{cMatrix}{a correlation matirx}

\item{gamma}{gamma value (default = 0.55)}

\item{t}{t value (default = 1)}

\item{lambda}{lambda value (default = 1)}
}
\value{
C
}
\description{
Author: Zhi Huang
}
