% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_lmQCM.R
\name{merging_lmQCM}
\alias{merging_lmQCM}
\title{merging_lmQCM: Subroutine for Merging Gene Clusters}
\usage{
merging_lmQCM(C, beta = 0.4, minClusterSize = 10)
}
\arguments{
\item{C}{Resulting clusters}

\item{beta}{beta value (default = 0.4)}

\item{minClusterSize}{minimum length of cluster to retain (default = 10)}
}
\value{
mergedCluster
}
\description{
Author: Zhi Huang
}
